Finished the basics of the getDellWarranty()
The basic purpose of the function is now done.
Added in was a for loop that will check the devices warranty, from there it will either state the latest warranty date or that there is no warranty for the device.
This is currently comparing basic warranty, later I will pull out specific warranty.

From here the "warrant" variable is pushed to an Array list and returned at the end of the function. This is so it can be used and outputted into the spreadsheet.