Creating weeklyMain()
This function will trigger from google sheets triggers set once a week.
It will wipe the whole spreadsheet and repopulate all the data. The reason behind doing this is because the organisation can extend warranty by purchasing more.
If you were to only ever update missing devices warranties. If more is purchased for an existing one it would then not change and accurately reflect it.